+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Removed
 * **BREAKING** Date Picker (Legacy): Remove ([#3262](https://github.com/dso-toolkit/dso-toolkit/issues/3262))
+* **BREAKING** Accordion: Remove HTML/CSS implementatie ([#3267](https://github.com/dso-toolkit/dso-toolkit/issues/3267))
 
 ## 🍟 Release 80.0.0 - 2025-10-30